requirements.txt
# ============================================
# AGENTE FISCAL - Dependências Python
# ============================================

# ==========================================
# WEB FRAMEWORK
# ==========================================
flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.0
blinker==1.9.0

# ==========================================
# PRODUCTION SERVER
# ==========================================
gunicorn==21.2.0

# ==========================================
# DATABASE
# ==========================================
sqlalchemy==2.0.35

# ==========================================
# SECURITY
# ==========================================
bcrypt==4.1.1
cryptography==46.0.3
cffi==2.0.0
pycparser==2.23

# ==========================================
# HTTP & API REQUESTS
# ==========================================
requests==2.31.0
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

# ==========================================
# AI - GOOGLE GEMINI
# ==========================================
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.62.3
proto-plus==1.26.1
protobuf==4.25.8
httplib2==0.31.0
uritemplate==4.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==6.2.1

# ==========================================
# DOCUMENT PROCESSING
# ==========================================
pdfplumber==0.10.3
pdfminer.six==20221105
pypdfium2==5.0.0
pillow==12.0.0

# ==========================================
# DATA VALIDATION
# ==========================================
pydantic==2.12.3
pydantic_core==2.41.4
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ==========================================
# UTILITIES
# ==========================================
python-dotenv==1.0.0
tqdm==4.67.1
packaging==25.0
pyparsing==3.2.5
colorama==0.4.6